moving AssetsMapping class to asset.py
