+++ lumibot/entities/__init__.py
@@ -1,5 +1,4 @@
-from .asset import Asset
-from .assets_mapping import AssetsMapping
+from .asset import Asset, AssetsMapping
 from .bar import Bar
 from .bars import Bars
 from .order import Order
