--- lumibot/entities/assets_mapping.py
@@ -1,27 +1,0 @@
-from collections import UserDict
-
-from .asset import Asset
-
-
-class AssetsMapping(UserDict):
-    def __init__(self, mapping):
-        UserDict.__init__(self, mapping)
-        symbols_mapping = {k.symbol: v for k, v in mapping.items()}
-        self._symbols_mapping = symbols_mapping
-
-    def __missing__(self, key):
-        if isinstance(key, str):
-            if key in self._symbols_mapping:
-                return self._symbols_mapping[key]
-        raise KeyError(key)
-
-    def __contains__(self, key):
-        if isinstance(key, str):
-            return key in self._symbols_mapping
-        return key in self.data
-
-    def __setitem__(self, key, value):
-        if isinstance(key, str):
-            self.data[Asset(key)] = value
-        else:
-            self.data[key] = value
